add frozen account feature

+++ app/Http/Controllers/API/Auth/LoginController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\API\Auth;
 
+use App\Exceptions\Auth\AccountIsFrozen;
 use App\Exceptions\Auth\InvalidCredentials;
 use App\Http\Controllers\API\Controller;
 use App\Http\Requests\API\Auth\LoginRequest;
@@ -15,6 +16,7 @@ class LoginController extends Controller
      * @group Auth
      * @responseFile app/Http/Responses/Samples/Auth/login.json
      * @responseFile 401 app/Http/Responses/Samples/Auth/invalid-credentials-exception.json
+     * @responseFile 401 app/Http/Responses/Samples/Auth/account-frozen-exception.json
      * @responseFile 422 app/Http/Responses/Samples/validation-error.json
      */
     public function __invoke(LoginRequest $request)
@@ -22,8 +24,14 @@ class LoginController extends Controller
         if(!Auth::guard('api')->attempt($request->validated()))
             throw new InvalidCredentials;
 
+        /** @var Student $user */
         $user = Student::where('phone' , $request->phone)->first();
+        if(!$user->canLogin())
+            throw new AccountIsFrozen;
+
         $user->tokens()->delete();
+        $user->incrementNumberOfLoginAttempts();
+
         return ['token' => $user->createToken('access-token')->plainTextToken];
     }
 }
